Now using fetch() and filters started
We now use the fetch api to issue a POST request to PHP which can then select the appropriate query and return the data in a JSON
This is easier for extending functionality.

We populate two list boxes with all unique values within the table for the given field.

There is a pre-programmed response to the APPLY button querying a specific product line. We can expand this later to use the options selected in the drop down.