Minor additions to bootstrap2 page
Nothing major, we just added the ability to select different styles for the demo table.
The colour of the table content and the table border is selected via 2 select dropdowns.
The changes are then applied with a button. A javascript file bootstrap2.js has been added
to support the changes.